bit of cleanup
