+++ python/fold_histos.py
@@ -23,24 +23,9 @@ inFile.cd("ana;1")
 keyList = inFile.GetKeyNames("ana")
 print "\nKeys in file:", len(keyList)
 
-def split_key(ll):
-    return map(lambda s: s.split('_'), ll)
-
-
-import pandas as pd
-def keys_to_df(lol):
-    uu = pd.DataFrame(lol, columns=['det','lay','U','V'])
-    # drop the last 6 rows which are spurious text
-    #print( uu.shape )
-    #print( uu.tail(5))
-    #print( uu.tail(5).index)
-    # print( uu.drop(uu.tail(5).index,inplace=True).shape )
-    # print( uu.drop(uu.tail(5).index,inplace=True).shape )
-    return uu[:-6]
-    # return  uu.drop(uu.tail(5).index,inplace=True)
 
 # print "\nKeys in file:", split_key(keyList)
-keys_df = keys_to_df( split_key(keyList) )
+keys_df = create_keys_df(keyList)
 
 print "\ndf head:", keys_df.head()
 print()
